+++ src/main/java/com/coinbase/prime/model/TransactionMetadata.java
@@ -20,6 +20,7 @@
 
 package com.coinbase.prime.model;
 import com.coinbase.prime.model.MatchMetadata;
+import com.coinbase.prime.model.RewardMetadata;
 import com.coinbase.prime.model.OnchainTransactionMetadata;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -35,12 +36,16 @@ public class TransactionMetadata {
     @JsonProperty("web3_transaction_metadata")
     private OnchainTransactionMetadata onchainTransactionMetadata;
 
+    @JsonProperty("reward_metadata")
+    private RewardMetadata rewardMetadata;
+
     public TransactionMetadata() {
     }
 
     public TransactionMetadata(Builder builder) {
         this.matchMetadata = builder.matchMetadata;
         this.onchainTransactionMetadata = builder.onchainTransactionMetadata;
+        this.rewardMetadata = builder.rewardMetadata;
     }
     public MatchMetadata getMatchMetadata() {
         return matchMetadata;
@@ -56,10 +61,19 @@ public class TransactionMetadata {
     public void setWeb3TransactionMetadata(OnchainTransactionMetadata onchainTransactionMetadata) {
         this.onchainTransactionMetadata = onchainTransactionMetadata;
     }
+    public RewardMetadata getRewardMetadata() {
+        return rewardMetadata;
+    }
+
+    public void setRewardMetadata(RewardMetadata rewardMetadata) {
+        this.rewardMetadata = rewardMetadata;
+    }
     public static class Builder {
         private MatchMetadata matchMetadata;
 
         private OnchainTransactionMetadata onchainTransactionMetadata;
+
+        private RewardMetadata rewardMetadata;
 
         public Builder matchMetadata(MatchMetadata matchMetadata) {
             this.matchMetadata = matchMetadata;
@@ -68,6 +82,11 @@ public class TransactionMetadata {
 
         public Builder onchainTransactionMetadata(OnchainTransactionMetadata onchainTransactionMetadata) {
             this.onchainTransactionMetadata = onchainTransactionMetadata;
+            return this;
+        }
+
+        public Builder rewardMetadata(RewardMetadata rewardMetadata) {
+            this.rewardMetadata = rewardMetadata;
             return this;
         }